// src/integrations/supabase/types.ts
export type Database = {
  // Allows to automatically instantiate createClient with right options
  // instead of createClient<Database, { PostgrestVersion: 'XX' }>(URL, KEY)
  __InternalSupabase: {
    PostgrestVersion: "14.1"
  }
  public: {
// ...
    Enums: {
      property_type:
        | "1 BHK"
        | "2 BHK"
        | "3 BHK"
        | "1 Bed"
        | "2 Bed"
        | "3 Bed"
        | "Studio"
      tenant_preference:
        | "Bachelor"
        | "Family"
        | "Girls Only"
        | "Working Professionals"
        | "Any"
      user_role: "customer" | "manager"
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}

type DatabaseWithoutInternals = Omit<Database, "__InternalSupabase">

type DefaultSchema = DatabaseWithoutInternals[Extract<keyof Database, "public">]
// ...
export type Enums<
  DefaultSchemaEnumNameOrOptions extends
    | keyof DefaultSchema["Enums"]
    | { schema: keyof DatabaseWithoutInternals },
  EnumName extends DefaultSchemaEnumNameOrOptions extends {
    schema: keyof DatabaseWithoutInternals
  }
    ? keyof DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"]
    : never = never,
> = DefaultSchemaEnumNameOrOptions extends {
  schema: keyof DatabaseWithoutInternals
}
  ? DatabaseWithoutInternals[DefaultSchemaEnumNameOrOptions["schema"]]["Enums"][EnumName]
  : DefaultSchemaEnumNameOrOptions extends keyof DefaultSchema["Enums"]
    ? DefaultSchema["Enums"][DefaultSchemaEnumNameOrOptions]
    : never
// ...
export const Constants = {
  public: {
    Enums: {
      property_type: [
        "1 BHK",
        "2 BHK",
        "3 BHK",
        "1 Bed",
        "2 Bed",
        "3 Bed",
        "Studio",
      ],
      tenant_preference: [
        "Bachelor",
        "Family",
        "Girls Only",
        "Working Professionals",
        "Any",
      ],
      user_role: ["customer", "manager"],
    },
  },
} as const